Add files via upload
The goal of the work is to understand the migration of planets in a disk. It is actually a norm for planets to migrate, Jupiter and Saturn have migrated too from somewhere within the Solar System. 
Midplane density of the 3D model of the protoplanetary disk (grid: 512 (radial) by 512 (azimuthal) by 64 (vertical)). 
240 planets are embedded in the disk, each planet has a unique color. A set of 30 planets at each of the 8 azimuthal locations has the same marker-symbol. Three annuli are resonance regions for spiral density waves in the disk.